rename band split folder to roformer_models
